Move blackbox out of API crate, since it's not shared
